requirements.txt
aiofiles==0.4.0
aiopg==0.16.0
httptools==0.0.13
Jinja2==2.10.1
multidict==4.5.2
psycopg2-binary==2.8.2
regex==2019.4.14
sanic==19.3.1
ujson==1.35
uvloop==0.12.2
websockets==6.0
